Wrap PIL.Image class in our own Image class
The pillow Image class is a bit cumbersome and confusing to use. Being
inside a module of the same name, we end up referring to PIL.Image.Image
everywhere.

Wrap that up in our own Image class. We'll provide an API that's a bit
nicer for our uses, and also hide the PIL Image class as an implementation
detail.