typing: improve mypy config for examples
Type checking is relaxed for the filter/pipeline examples, because they're
intended to present the scripting opportunities for the pipeline API in a
simple way without requiring type hints on everything.

It would still be nice to catch type errors where possible, so instead of
excluding the examples from type checking completely, choose some looser
requirements for them.